Created class for bouncing audio
Allows bounceAudio to be interrupted
