+++ common/src/main/kotlin/org/wycliffeassociates/otter/common/domain/narration/AudioFileUtils.kt
@@ -2,10 +2,6 @@ package org.wycliffeassociates.otter.common.domain.narration
 
 import org.wycliffeassociates.otter.common.audio.AudioFile
 import org.wycliffeassociates.otter.common.audio.AudioFileReader
-import org.wycliffeassociates.otter.common.audio.wav.WavFile
-import org.wycliffeassociates.otter.common.audio.wav.WavOutputStream
-import org.wycliffeassociates.otter.common.data.audio.AudioMarker
-import org.wycliffeassociates.otter.common.domain.audio.OratureAudioFile
 import org.wycliffeassociates.otter.common.persistence.IDirectoryProvider
 import java.io.File
 import javax.inject.Inject
@@ -41,32 +37,4 @@ class AudioFileUtils @Inject constructor(private val directoryProvider: IDirecto
             }
         }
     }
-}
-
-/**
- * Constructs a Wav file at path specified by bouncedAudio with audio data provided by the reader parameter and cues
- * provided by the markers parameter
- */
-fun bounceAudio(bouncedAudio: File, reader: AudioFileReader, markers: List<AudioMarker>) {
-    val bytes = ByteArray(DEFAULT_BUFFER_SIZE)
-
-    reader.open()
-    reader.seek(0)
-
-    if (bouncedAudio.exists() && bouncedAudio.length() > 0) {
-        bouncedAudio.delete()
-    }
-    val wav = WavFile(bouncedAudio, 1, 44100, 16)
-    WavOutputStream(wav).use { out ->
-        while (reader.hasRemaining()) {
-            val read = reader.getPcmBuffer(bytes)
-            out.write(bytes, 0, read)
-        }
-    }
-    wav.update()
-    val oaf = OratureAudioFile(bouncedAudio)
-    for (verse in markers) {
-        oaf.addMarker<AudioMarker>(verse.clone())
-    }
-    oaf.update()
 }
